Refactor: 패키지명을 auth에서 authentication으로 변경
Close #37
