+++ src/main/java/com/uranus/taskmanager/api/authentication/AuthenticationInterceptor.java
@@ -1,10 +1,10 @@
-package com.uranus.taskmanager.api.auth;
+package com.uranus.taskmanager.api.authentication;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import com.uranus.taskmanager.api.auth.exception.UserNotLoggedInException;
+import com.uranus.taskmanager.api.authentication.exception.UserNotLoggedInException;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
