+++ src/main/java/com/uranus/taskmanager/api/authentication/LoginMember.java
@@ -1,11 +1,11 @@
-package com.uranus.taskmanager.api.auth;
+package com.uranus.taskmanager.api.authentication;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginRequired {
+public @interface LoginMember {
 }
