+++ src/main/java/com/uranus/taskmanager/api/authentication/SessionKey.java
@@ -1,4 +1,4 @@
-package com.uranus.taskmanager.api.auth;
+package com.uranus.taskmanager.api.authentication;
 
 public abstract class SessionKey {
 	public static final String LOGIN_MEMBER = "loginId";
