+++ src/main/java/com/uranus/taskmanager/api/authentication/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package com.uranus.taskmanager.api.auth.exception;
+package com.uranus.taskmanager.api.authentication.exception;
 
 import org.springframework.http.HttpStatus;
