+++ src/main/java/com/uranus/taskmanager/api/authentication/service/AuthenticationService.java
@@ -1,12 +1,12 @@
-package com.uranus.taskmanager.api.auth.service;
+package com.uranus.taskmanager.api.authentication.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.uranus.taskmanager.api.auth.dto.request.LoginRequest;
-import com.uranus.taskmanager.api.auth.dto.response.LoginResponse;
-import com.uranus.taskmanager.api.auth.exception.InvalidLoginIdentityException;
-import com.uranus.taskmanager.api.auth.exception.InvalidLoginPasswordException;
+import com.uranus.taskmanager.api.authentication.dto.request.LoginRequest;
+import com.uranus.taskmanager.api.authentication.dto.response.LoginResponse;
+import com.uranus.taskmanager.api.authentication.exception.InvalidLoginIdentityException;
+import com.uranus.taskmanager.api.authentication.exception.InvalidLoginPasswordException;
 import com.uranus.taskmanager.api.member.domain.Member;
 import com.uranus.taskmanager.api.member.repository.MemberRepository;
 import com.uranus.taskmanager.api.security.PasswordEncoder;
