+++ src/main/java/com/uranus/taskmanager/api/workspace/service/WorkspaceCreateService.java
@@ -1,6 +1,6 @@
 package com.uranus.taskmanager.api.workspace.service;
 
-import com.uranus.taskmanager.api.auth.dto.request.LoginMemberDto;
+import com.uranus.taskmanager.api.authentication.dto.request.LoginMemberDto;
 import com.uranus.taskmanager.api.workspace.dto.request.WorkspaceCreateRequest;
 import com.uranus.taskmanager.api.workspace.dto.response.WorkspaceResponse;
