+++ src/test/java/com/uranus/taskmanager/fixture/RestAssuredAuthenticationFixture.java
@@ -3,7 +3,7 @@ package com.uranus.taskmanager.fixture;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 
-import com.uranus.taskmanager.api.auth.dto.request.LoginRequest;
+import com.uranus.taskmanager.api.authentication.dto.request.LoginRequest;
 import com.uranus.taskmanager.api.member.dto.request.SignupRequest;
 
 import io.restassured.RestAssured;
