+++ src/test/java/com/uranus/taskmanager/fixture/TestFixture.java
@@ -1,6 +1,6 @@
 package com.uranus.taskmanager.fixture;
 
-import com.uranus.taskmanager.api.auth.dto.request.LoginMemberDto;
+import com.uranus.taskmanager.api.authentication.dto.request.LoginMemberDto;
 import com.uranus.taskmanager.api.invitation.InvitationStatus;
 import com.uranus.taskmanager.api.invitation.domain.Invitation;
 import com.uranus.taskmanager.api.member.domain.Member;
